requirements.txt
streamlit
pandas
PyGithub
pyyaml
plotly